rename ColorLegendOrdinalHorizontal to ColorLegendOrdinalChips
