+++ packages/ui/src/lib/colorLegends/ColorLegendOrdinalChips.svelte
